pyproject.toml
[project]
name = "hewo-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "google-api-python-client>=2.187.0",
    "google-auth-httplib2>=0.2.1",
    "google-auth-oauthlib>=1.2.3",
    "nextcord[speed]>=3.1.1",
    "pydantic-settings>=2.12.0",
    "pymysql>=1.1.2",
    "requests>=2.32.5",
    "sqlmodel>=0.0.27",
]

[project.scripts]
main = "main:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
